+++ okdata/cli/commands/datasets/boilerplate/__init__.py
@@ -1,3 +1,0 @@
-from .boilerplate import DatasetsBoilerplateCommand
-
-__all__ = ["DatasetsBoilerplateCommand"]